+++ src/util/patcher/channelTextAreaButtons.js
@@ -42,7 +42,7 @@ export const patch = (tooltipText, imgSrc, clickHandler, { inUpload = false, inR
             onMouseEnter,
             onMouseLeave
           },
-            React.createElement("img", {
+            typeof imgSrc !== 'string' ? imgSrc : React.createElement("img", {
               src: imgSrc,
               width: "24px",
               height: "24px",