+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
   - ### Patcher API Improvements
     - Added error handling if patches cause an error (it now just skips the patch instead of likely crashing)
-    - [GitHub Issue #18](https://github.com/GooseMod/Injector/issues/18): channelTextAreaButtons.patch: Added optional options to show button in upload modal and in readonly channels (now both not showing by default)
+    - [GitHub Issue #18](https://github.com/GooseMod/Injector/issues/18): ChannelTextAreaButtons: Added optional options to show button in upload modal and in readonly channels (now both not showing by default)
 
   - ### DOM To React
     - [GitHub Issue #26](https://github.com/GooseMod/Injector/issues/26): Changed goosemod.confirmDialog to use React instead of DOM